Clean up build script.

+++ build.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
-rm -r cookbooks
+if [ -d "$DIRECTORY" ]; then
+	rm -r cookbooks
+fi
+
+bundle install
 bundle exec berks vendor cookbooks
 packer build packer.json
